pyproject.toml
[tool.poetry]
name = "nndebugger"
version = "0.1.1"
description = ""
authors = ["Rishi Gurnani <rgurnani96@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.3,<=3.7.5"
numpy = "^1.17"

[tool.poetry.dev-dependencies]
torch = { url = "https://download.pytorch.org/whl/cu102/torch-1.9.1%2Bcu102-cp37-cp37m-linux_x86_64.whl" }
torch-scatter = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu102/torch_scatter-2.0.9-cp37-cp37m-linux_x86_64.whl"}
torch-sparse = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu102/torch_sparse-0.6.12-cp37-cp37m-linux_x86_64.whl"}
torch-geometric = "^2.0.2"
rdkit-pypi = "^2021.9.2"
pytest = "^6.2.5"
ipykernel = "^6.6.0"
black = "^21.12b0"
# mpnn-v2 = {git = "ssh://git@github.com/rishigurnani/mpnn_v2.git", rev = "0.1.10"}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
